Warewolf Server now handles a GraphQL request when using .gql
Updated the Warewolf Server to handle a .gql request as a GraphQL request.
Updated the System.Reactive libraries to the latest version, should bring a slight performance and stability improvement.
The GraphQL query is run against the "Output" version of an environment for that execution i.e. it will not return a variable that has not been marked as Output.